Task 3a
For this we have generated a json with ICD-10 condition codes and posted it in Local HAPI server UI  in the CodeSystem resource. We have taken “code” and “display” information from the ICD-10 url:https://icd.who.int/browse10/2019/en#/ and populated it in a json. Please find the json as :  ICD_10Conditions.json.